+++ stwcs/updatewcs/utils.py
@@ -1,4 +1,5 @@
 import os
+import string
 from stsci.tools import fileutil
 
 import logging